Fixed compilation errors and warnings with MinGW and MSVC.

+++ src/libboardgame_base/Transform.cpp
@@ -176,8 +176,8 @@ unsigned int TransfTrigonRefl::get_new_point_type() const
 
 CoordPoint TransfTrigonRot60::get_transformed(const CoordPoint& p) const
 {
-    float x = floor(0.5 * p.x + 1.5 * p.y);
-    float y = floor(-0.5 * p.x + 0.5 * p.y);
+    int x = static_cast<int>(floor(0.5f * p.x + 1.5f * p.y));
+    int y = static_cast<int>(floor(-0.5f * p.x + 0.5f * p.y));
     return CoordPoint(x, y);
 }
 
@@ -195,8 +195,8 @@ unsigned int TransfTrigonRot60::get_new_point_type() const
 
 CoordPoint TransfTrigonRot120::get_transformed(const CoordPoint& p) const
 {
-    float x = floor(-0.5f * p.x + 1.5f * p.y);
-    float y = ceil(-0.5f * p.x - 0.5f * p.y);
+    int x = static_cast<int>(floor(-0.5f * p.x + 1.5f * p.y));
+    int y = static_cast<int>(ceil(-0.5f * p.x - 0.5f * p.y));
     return CoordPoint(x, y);
 }
 
@@ -231,8 +231,8 @@ unsigned int TransfTrigonRot180::get_new_point_type() const
 
 CoordPoint TransfTrigonRot240::get_transformed(const CoordPoint& p) const
 {
-    float x = ceil(-0.5f * p.x - 1.5f * p.y);
-    float y = ceil(0.5f * p.x - 0.5f * p.y);
+    int x = static_cast<int>(ceil(-0.5f * p.x - 1.5f * p.y));
+    int y = static_cast<int>(ceil(0.5f * p.x - 0.5f * p.y));
     return CoordPoint(x, y);
 }
 
@@ -250,8 +250,8 @@ unsigned int TransfTrigonRot240::get_new_point_type() const
 
 CoordPoint TransfTrigonRot300::get_transformed(const CoordPoint& p) const
 {
-    float x = ceil(0.5f * p.x - 1.5f * p.y);
-    float y = floor(0.5f * p.x + 0.5f * p.y);
+    int x = static_cast<int>(ceil(0.5f * p.x - 1.5f * p.y));
+    int y = static_cast<int>(floor(0.5f * p.x + 0.5f * p.y));
     return CoordPoint(x, y);
 }
 
@@ -269,8 +269,8 @@ unsigned int TransfTrigonRot300::get_new_point_type() const
 
 CoordPoint TransfTrigonReflRot60::get_transformed(const CoordPoint& p) const
 {
-    float x = floor(0.5 * (-p.x) + 1.5 * p.y);
-    float y = floor(-0.5 * (-p.x) + 0.5 * p.y);
+    int x = static_cast<int>(floor(0.5f * (-p.x) + 1.5f * p.y));
+    int y = static_cast<int>(floor(-0.5f * (-p.x) + 0.5f * p.y));
     return CoordPoint(x, y);
 }
 
@@ -288,8 +288,8 @@ unsigned int TransfTrigonReflRot60::get_new_point_type() const
 
 CoordPoint TransfTrigonReflRot120::get_transformed(const CoordPoint& p) const
 {
-    float x = floor(-0.5f * (-p.x) + 1.5f * p.y);
-    float y = ceil(-0.5f * (-p.x) - 0.5f * p.y);
+    int x = static_cast<int>(floor(-0.5f * (-p.x) + 1.5f * p.y));
+    int y = static_cast<int>(ceil(-0.5f * (-p.x) - 0.5f * p.y));
     return CoordPoint(x, y);
 }
 
@@ -324,8 +324,8 @@ unsigned int TransfTrigonReflRot180::get_new_point_type() const
 
 CoordPoint TransfTrigonReflRot240::get_transformed(const CoordPoint& p) const
 {
-    float x = ceil(-0.5f * (-p.x) - 1.5f * p.y);
-    float y = ceil(0.5f * (-p.x) - 0.5f * p.y);
+    int x = static_cast<int>(ceil(-0.5f * (-p.x) - 1.5f * p.y));
+    int y = static_cast<int>(ceil(0.5f * (-p.x) - 0.5f * p.y));
     return CoordPoint(x, y);
 }
 
@@ -343,8 +343,8 @@ unsigned int TransfTrigonReflRot240::get_new_point_type() const
 
 CoordPoint TransfTrigonReflRot300::get_transformed(const CoordPoint& p) const
 {
-    float x = ceil(0.5f * (-p.x) - 1.5f * p.y);
-    float y = floor(0.5f * (-p.x) + 0.5f * p.y);
+    int x = static_cast<int>(ceil(0.5f * (-p.x) - 1.5f * p.y));
+    int y = static_cast<int>(floor(0.5f * (-p.x) + 0.5f * p.y));
     return CoordPoint(x, y);
 }
